Update bl2 storage load func, add pll init
Conflicts:

	arch/arm/cpu/armv8/common/firmware/common/fip.c
	arch/arm/cpu/armv8/common/firmware/plat/gxb/storage.c

+++ arch/arm/cpu/armv8/common/firmware/plat/gxb/include/storage.h
@@ -28,11 +28,11 @@
 #define MAX_DESC_NUM                    8
 #define MAX_BLOCK_COUNTS                512
 
-uint32_t storage_load(uint32_t src, uint32_t des, uint32_t size, const char * image_name);
-uint32_t get_boot_device(void);
-uint32_t spi_read(uint32_t src, uint32_t des, uint32_t size);
-unsigned sdio_read_blocks(struct sd_emmc_global_regs *sd_emmc_regs, uint32_t src, uint32_t des, uint32_t size,uint32_t mode);
-unsigned sdio_read_data(unsigned boot_device, uint32_t src, uint32_t des, uint32_t size);
+uint64_t storage_load(uint64_t src, uint64_t des, uint64_t size, const char * image_name);
+uint64_t get_boot_device(void);
+uint64_t spi_read(uint64_t src, uint64_t des, uint64_t size);
+uint64_t sdio_read_blocks(struct sd_emmc_global_regs *sd_emmc_regs, uint64_t src, uint64_t des, uint64_t size,uint64_t mode);
+uint64_t sdio_read_data(uint64_t boot_device, uint64_t src, uint64_t des, uint64_t size);
 
 /*SIZE defines*/
 #define SIZE_1K							0x400
